Load LiveDisplay classes on-demand so we don't need X if only using eg the PNG backend.
Return undefined data bounds correctly now that this is possible.

Only allow Painter.render () on parentless painters.

Stroke a line every 100 points in XYPainter so that the path
doesn't get absurdly big with a lot of data.

Don't write the PageSize DSC features since we may not always
be Letter size.



git-svn-id: https://newton.cx/svn/trunk/omegaplot@1017 69fc08fc-fdfc-0310-ace8-9e8da0474e22