I initially set out to make LiveDisplays work more like they should, where the underlying OmegaArea (optionally) holds a weak ref to the painter, and stops painting it if the underlying painter disappears. This emulates the old behavior where del'ing the right thing makes the plot go away.
But as it turns out that was foiled because children of a
painter hold a reference to their parents.

But, I never actually use that parent info, and I never liked
that code, so I just ripped it all out. Probably later I'll
have to add something back in but things seem to be just fine
for now.

Also, update to renamed Cairo python binding functions.



git-svn-id: https://newton.cx/svn/trunk/omegaplot@1079 69fc08fc-fdfc-0310-ace8-9e8da0474e22