Reinsert the parent-tracking code, with a few tweaks. It makes me unhappy but seems to be OK.
Note that GTK won't let you do very complicated things with parenting
at all. Eg, to reparent a widget, you have to manually remove it from
its old container, manually remove an existing child in the destination
(if necessary), and then put it inside the new container. Maybe 
disallowing fancy ops that we attempt is the way to go.



git-svn-id: https://newton.cx/svn/trunk/omegaplot@1083 69fc08fc-fdfc-0310-ace8-9e8da0474e22