Add a very basic sin(x) demo to help with testing.
Clean up the DiscreteAxis stuff -- it is silly to do
all the mapping to integers ourselves. May break histograms
and things, haven't checked.

Futz around with the style application to get stampers to 
use  colors correctly. Need to figure out this stuff 
for real at some point.

Tweak the colors so that a plot with only two elements
doesn't look so ugly.



git-svn-id: https://newton.cx/svn/trunk/omegaplot@1138 69fc08fc-fdfc-0310-ace8-9e8da0474e22