Fix failure to draw final ticks/labels on nudged axes.
From the code:

NOTE: 'inc' can fall prey to floating-point inexactness, e.g. "0.2"
really is 0.2 + 1.1e-17. This can be a problem when the bounds have
been nudged to be nice and round because the inexactness may
accumulate as we increment 'val', resulting in the final tick not
being drawn and labeled because 'val' is 6 + 1e-16, not 6 exactly. To
combat this, we round off 'val' when it is at major tick, which have
nice round values.