Add LogValueAxisPainter and incorporate it into setLinLogAxes.
The LogarithmicAxisPainter becomes not very useful when you want to
plot the log value of some variable but the variable doesn't span a
large range, since there are so few labels. For instance, plotting log
values between 1e3 and 1e3+1 fails pretty badly. Add the
LogValueAxisPainter, which labels the axis in a linear style with the
logs of the data values. In the above example, it would show a linear
axis ranging from 3 to 3.0004, for instance. This makes it possible
to actually read values off of the plot while still plotting in log
space. If this is done, the axis labels must be changed to make it
clear that log values are being plotted, but that isn't something we
can do in software.

Also change setLinLogAxes to be able to use a LogValueAxisPainter
instead of the usual LogarithmicAxisPainter depending on a defaulted
keyword argument.