rect: fix occasional bad bounds in RectPlot.rebound ()
RectPlot.rebound () first called field.setBounds() using the bounds of
the first FieldPainter, then field.expandBounds() using the bounds of
all the subsequent FieldPainters. This would break if the first
FieldPainter didn't define all of its bounds (e.g., if it was a VLine)
and the final bounds ought not have contained the default (0-10, 0-10)
boundary region. Change it to calculate things in a way such that the
default bounds will be overridden correctly when appropriate.