latex: add non-beautiful helpers for scaling LaTeX output
I often find myself writing quick hacks to scale the output of
LaTeX renderings to tweak the size of the output -- the default
dvips flags have worked well in most circumstances, but
sometimes they need adjustment. Add some hooks to make it easier
to do this scaling. There should really be some nice generic
way of figuring out the appropriate text size, but that hasn't
been forthcoming, and it is definitely helpful to be able to
fudge things a bit more easily.