omega/stamps.py: stroke *and* fill when drawing filled stamps
The stroke line around a stamp extends outside of the path coordinates by
0.5*linewidth, while the fill fills precisely inside the path coordinates.
Therefore, if we draw a filled stamp merely by filling the desired path, the
unfilled version of the stamp will be slightly larger. This becomes
distracting if we're drawing both filled and unfilled stamps and the linewidth
is nonnegligible.

Fix this by always stroking. Now filled and unfilled stamps have identical
outer edges. This looks much better.