omega/styles.py: remove 'thickLine' as a style parameter
Instead put everything in terms of fineLine. Also set the linesize upon
initialization of the context, and don't override it in other places.

This changes the default look, which used to have data lines having
width=thickLine and axis tick lines having width=fineLine. So far I think this
change is beneficial.