IPython integration and background GTK+ looping were totally busted.
The IPython API changed a while ago, and it has a new system for running
graphics mainloops. We can now detect IPython again, and if the user runs
"%gui gtk" we will notice this and DTRT.

The GTK stuff is somewhat sketchy since it relies on IPython to know if the
mainloop is running somewhere. But such is life.

Also, since omega.render caches display pagers, things will break easily if
you actually use the %gui magic to toggle GTK support on and off in a live
IPython session. We should merge the {Yes,No}LoopDisplayPagers to make
game-time decisions, but that's more work than I feel like doing right now.