omega/rect.py: fix a bug I just introduced
It appears that ctxt.restore() doesn't reset the path the way I thought it
did, so we need to make sure that it's empty when we finish drawing an
{X|Y}Band. In particular, if we filled but didn't stroke, we'd leave the path
active. An alternate approach would be to only fill_preserve if we're both
filling and stroking; it doesn't seem to me that either approach is
particularly superior.