New high-level function omega.latest()
This gets you the Painter that was most recently shown in a display pager.
This is helpful since I often just type:

>>> om.quickXY (x, y).show ()
>>> ...

And then realize that I want to add more doodads to the plot. One can recover
the pager in IPython with the output history variables ("_3", etc.) but that's
kind of a hassle and I think I'll like writing

>>> p = om.latest ()

more.