Clean up stamp stroke/fill choices
This was always kind of awkwardly implemented, and there was a particular
limitation where you couldn't stroke and fill a stamp using different styles,
which means you couldn't easily outline a stamp for emphasis or things like
that. I've never been happy with this code so I've gone and tried to
standardize it all in a way that should be more flexible and Just Work.

The changes to the "symfunc" code in styles.py are untested. I'm pretty sure
they're OK. Unfortunately the MultiStamp is hairy enough where it feels like a
pain to try to do a quick test of it.