Update for background rendering in the Jupyter Console
We need yet another alternate method for detecting whether the Gtk mainloop is
running. I've renamed the related bits from IPython-y things to Jupyter-y things,
but I plan to still maintain support for IPython as needed. It's just true that
things will transition to Jupyter more and more as time goes on.