omega/jupyter.py: fix for IPython 5.0
Another version of IPython, another way that we need to check if the Gtk
mainloop is running. Also a good time to update for my newfound discovery that
IPython is not on its way out.

I'm still surprised that I can't figure out a way to test if the mainloop is
running using only tools provided by `gi.repository.{GLib,Gtk}`, but that
remains the case.