Revert "Start preferring the cairocffi module to cairo"
This reverts commit 957db49c2a74da47579184c5ae31fd7040b93595.

I had thought that cairocffi was the answer to my problems with pycairo, but
it turns out that it is not currently compatible with the pygobject system,
which is a very important use case for me. In an amazing coincidence, just a
few hours ago, someone took charge and made a new release of pycairo that
fixes its problems. So, back to that.