omega/rect.py: finally fix the infinite-loop bug
It could happen that certain value ranges were *just* above the floating-point
precision threshold, such that we calculated that the range was non-zero in size,
but then if we tried to subdivide it for placing axis labels, we'd have a case of
`val + inc = val` and we'd loop infinitely. Fix this.